+++ src/index.ts
@@ -2,3 +2,6 @@ export * from '../packages/harness-core/src/index';
 export * from '../packages/harness-browser/src/index';
 export * from '../packages/harness-python/src/index';
 export * from '../packages/harness-javascript/src/index';
+export * from '../packages/harness-java/src/index';
+export * from '../packages/harness-csharp/src/index';
+export * from '../packages/harness-cpp/src/index';